refactor(extension): Share less.
